+++ src/entity/types.ts
@@ -1,0 +1,13 @@
+export interface PlaceType {
+  latitude: number,
+  longitude: number,
+  placeName: string,
+  distance: string,
+}
+
+export interface ScreenType {
+  isVisibleMap: boolean,
+  isVisibleList: boolean,
+  isVisibleClosestToilets: boolean,
+  isVisibleLandingScreen: boolean,
+}
